+++ packages/core/src/sdk/clients/beekeeper.ts
@@ -1,13 +1,11 @@
 import { createClient } from "@connectrpc/connect";
 import { App, BeekeeperService } from "../../gen";
-import { makeSingletonFactory } from "../utilities/utils";
-import { ClientOptions } from "./types";
-import { createTransport, invalidateHiveToken } from "./utils";
+import { getInterceptors } from "./token-manager";
+import { createTransport } from "./utils";
 
-export const createSingletonBeekeeperClient = makeSingletonFactory((options: ClientOptions<any>) => {
-  const transport = createTransport(options, App.BEEKEEPER);
+export const createBeekeeperClient = () => {
+  const transport = createTransport(App.BEEKEEPER, getInterceptors());
   return {
     ...createClient(BeekeeperService, transport),
-    invalidateHiveToken: invalidateHiveToken,
   };
-});
+};