+++ packages/core/src/sdk/repo/globals.ts
@@ -1,17 +1,10 @@
-import { VespaDatabaseStack } from "../../gen";
+import { UserType, VespaDatabaseStack } from "../../gen";
+import { ClientOptions } from "../clients";
+import { initialize } from "../clients/globals";
 import { getEnvString } from "../utilities";
 
-let accessToken = "";
 let stackHRN = "";
 export let databaseStack: VespaDatabaseStack | undefined = undefined;
-
-export const setAccessToken = (token: string) => {
-  accessToken = token;
-};
-
-export const getAccessToken = () => {
-  return accessToken;
-};
 
 export const setStackHRN = (hrn: string) => {
   stackHRN = hrn;
@@ -21,9 +14,13 @@ export const getStackHRN = () => {
   return stackHRN;
 };
 
-export const vespaInit = (options?: { stackHRN: string; accessToken: string }): void => {
-  setAccessToken(options?.accessToken || getEnvString("HIVE_ACCESS_TOKEN"));
-  setStackHRN(options?.stackHRN || getEnvString("HIVE_STACK_HRN"));
+export const vespaInit = (options: { stackHRN?: string; accessToken?: string; userType?: UserType; clientOptions: ClientOptions }): void => {
+  stackHRN = options.stackHRN || getEnvString("HIVE_STACK_HRN");
+  initialize({
+    token: options.accessToken || getEnvString("HIVE_ACCESS_TOKEN"),
+    type: options.userType || UserType.TENANT_SECURE_APP,
+    options: options.clientOptions,
+  });
 };
 
 export const setDatabaseStack = (stack: VespaDatabaseStack) => {