+++ packages/core/src/sdk/utilities/constants.ts
@@ -1,5 +1,4 @@
-import { App, Framework } from "../../gen";
-import { ClientType } from "./types";
+import { App, Framework, JavaScriptClientType } from "../../gen";
 
 export const DOMAIN = "hiveops.io";
 
@@ -16,8 +15,8 @@ export const APP_MAP: { [key in keyof typeof App]: string } = {
   VESPA: "vespa",
 };
 
-export const CLIENT_TYPE_FRAMEWORK_MAP: { [key in ClientType]: Framework } = {
-  node: Framework.GRPC,
-  deno: Framework.GRPC,
-  web: Framework.GRPC_WEB,
+export const CLIENT_TYPE_FRAMEWORK_MAP: { [key in keyof typeof JavaScriptClientType]: Framework } = {
+  UNSPECIFIED: Framework.UNSPECIFIED,
+  NODE: Framework.GRPC,
+  WEB: Framework.GRPC_WEB,
 };