+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @hiveops/cli
 
+## 1.3.0
+
+### Minor Changes
+
+- feat: enhance applyMigration function with async handling
+
+## 1.2.0
+
+### Minor Changes
+
+- feat: refactor client creation methods and improve token management; remove deprecated client options
+
+### Patch Changes
+
+- Updated dependencies
+  - @hiveops/core@1.6.0
+  - @hiveops/node@1.4.0
+
+## 1.1.0
+
+### Minor Changes
+
+- Refactor SDK clients to enhance token management and interceptor usage
+
+### Patch Changes
+
+- Updated dependencies
+  - @hiveops/core@1.5.0
+  - @hiveops/node@1.3.0
+
 ## 1.0.0
 
 ### Major Changes