+++ packages/core/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @hiveops/core
 
+## 1.7.1
+
+### Patch Changes
+
+- no-op
+
+## 1.7.0
+
+### Minor Changes
+
+- feat: update vespaInit function to support optional parameters for stackHRN, accessToken, and userType
+
+## 1.6.0
+
+### Minor Changes
+
+- feat: refactor client creation methods and improve token management; remove deprecated client options
+
+## 1.5.2
+
+### Patch Changes
+
+- feat: integrate JavaScriptClientType enum into client options and utility functions
+
+## 1.5.1
+
+### Patch Changes
+
+- actor: reorganize imports and enhance token management in SDK clients
+
+## 1.5.0
+
+### Minor Changes
+
+- Refactor SDK clients to enhance token management and interceptor usage
+
 ## 1.4.0
 
 ### Minor Changes