+++ packages/node/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @hiveops/node
 
+## 1.5.0
+
+### Minor Changes
+
+- feat: update vespaInit function to support optional parameters for stackHRN, accessToken, and userType
+
+### Patch Changes
+
+- Updated dependencies
+  - @hiveops/core@1.7.0
+
+## 1.4.0
+
+### Minor Changes
+
+- feat: refactor client creation methods and improve token management; remove deprecated client options
+
+### Patch Changes
+
+- Updated dependencies
+  - @hiveops/core@1.6.0
+
+## 1.3.1
+
+### Patch Changes
+
+- feat: update clientType in clientOptions to use JavaScriptClientType enum
+
+## 1.3.0
+
+### Minor Changes
+
+- Refactor SDK clients to enhance token management and interceptor usage
+
+### Patch Changes
+
+- Updated dependencies
+  - @hiveops/core@1.5.0
+
 ## 1.2.0
 
 ### Minor Changes